Force the rttest statistics to be fixed.
The regex that we use over in demos to test some of this
stuff out only handles floating-point numbers, not scientific
notation.  The mean latency, in particular, can grow into
a value large enough that it will get put into scientific
notation by default.  This fixes this output to be fixed
so the regex will always be able to find it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>